+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.155 (2023-12-5)
 
-- chore: save project info to git
+- feat: new projects pages
 
 ## 0.1.154 (2023-12-1)